# app/src/swagger.yaml
swagger: "2.0"
info:
    version: 1.0.0
    title: API
    description: Contele Teste :)
consumes:
    - application/json
produces:
    - application/json
host: "localhost:4000"

paths:
    /ping:
        x-swagger-router-controller: ping
        get:
            summary: health check :)
            description: Check api health
            operationId: ping
            tags:
                - ping
            responses:
                "200":
                    description: API OK YAY
                    schema:
                        type: object
                        properties:
                            message:
                                type: string
    /pois:
        x-swagger-router-controller: pois
        get:
            summary: get pois
            description: get poi by namee
            operationId: getPois
            tags:
                - pois
            parameters:
                - in: query
                  name: poiName
                  type: string
                - in: query
                  name: limit
                  type: integer
                  minimum: 1
                  maximum: 50
                - in: query
                  name: offset
                  type: integer
                  minimum: 0
            responses:
                "200":
                    description: success
                    schema:
                        type: array
                        items:
                            $ref: "#/definitions/Marker"
                "500":
                    description: failed
                    schema:
                        $ref: "#/definitions/ErrorResponse"
definitions:
    Marker:
        type: object
        additionalProperties: false
        properties:
            type: 
                type: string
                example: "Feature"
            properties:
                type: object
                additionalProperties: false
                properties:
                    cluster: 
                        type: boolean
                    poiName: 
                        type: string
                        example: "Localzinho"
            geometry:
                type: object
                additionalProperties: false
                properties:
                    type: 
                        type: string
                        example: "Point"
                    coordinates: 
                        type: array
                        items:
                            type: number
                        minItems: 2
                        maxItems: 2
                        example: 
                            items:
                            - -23.31509536897005
                            - -46.57099951314262
    ErrorResponse:
        type: object
        additionalProperties: false
        properties:
            error: 
                type: string
            message:
                type: string
                        